added and amend Junit for Find, Card and Transfer
